[#25718]YSQL: Fix memory leaks and segmentation faults in YSQL Connection Manager
Summary:
This patch helps resolve multiple segmentation faults and memory leaks that occur in stress test scenarios with Connection Manager.

- hashmap.c: Properly free the hashmap if allocation fails
- cron.c: Add length check on user name (already existed for DB)
- instance.c: If shared memory operations failed, we would return NULL from the wrapper function `yb_get_stats_ptr`. This is incorrect because shared memory operations fail with `(void*) -1`. Also added a missing `goto error` for this error scenario.
- storage.c: Properly free storage object if allocation fails

Note that some changes are speculative fixes for which we are waiting on QA for additional verification before closing individual tickets; the Jira ticket linked to this patch points to a Story that tracks the closure of core dumps that occur in stress test scenarios.

Jira: DB-14991

Test Plan: Jenkins: enable connection manager, all tests

Reviewers: skumar, mkumar, vpatibandla, devansh.saxena

Reviewed By: devansh.saxena

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D41483