[#25752] CDC: Stop sending cleanup rpc to master for logical replication
Summary:
Update Peers and Metrics thread periodically reads the state table and identifies expired / not of interest tables. A list of such tables is sent to the master in an rpc. Master than performs a cleanup operation, moving these tables to unqualified list and setting their checkpoint to max. However this cleanup is performed only for tables under gRPC replication. The list sent by update peers and metrics can contain tables under logical replication.

This diff introduces changes to not add the expired / not of interest table to the list to be sent to the master if they belong to logical replication.
Jira: DB-15042

Test Plan: Existing tests

Reviewers: skumar, siddharth.shah

Reviewed By: siddharth.shah

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D41493