[PLAT-16597][dr] A dropped table from source while in replication is reported twice once with source table id and once with target table id
Summary:
Previously, if the user dropped a table from the source universe without removing it from the replication first, it stayed in the cluster config and also it was added again to the xCluster config as a table in a transient status and thus two tables with different table ids (one on the source and one on the target) shows up in 'Dropped From Source' status.

This diff avoid adding the table again if the source table id is already part of the cluster config.

Test Plan: Create a dr config and dropped a source table and made sure it is reported once with the correct table id.

Reviewers: cwang, vbansal, nbhatia, amindrov

Reviewed By: amindrov

Subscribers: svc_phabricator, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41498