[PLAT-16383] K8S operator Storage config and Release to use secrets
Summary:
The operator previously took tokens, passwords, and other secrets as plain text directly
in the storage config and release CRDs. This updates them to use secrets instead.
The old method is not deprecated, but if a secret is provided it is used instead of
the plain text secret

Test Plan: created storage config and release with secrets

Reviewers: anijhawan, vkumar

Reviewed By: anijhawan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41525