[#25743] YSQL: Ignore collversion for en_US collations in TestYsqlUpgrade
Summary:
The test `TestYsqlUpgrade#migratingIsEquivalentToReinitdb` checks that the system catalog of a newly-created cluster is in an identical state to that of an old cluster which has been upgraded with `ysql_upgrade`.

When there are changes that add a system object (e.g. a system function), this function needs a row in the pg_proc system table. When you add a system function, you add the corresponding row in `pg_proc.dat`, which gets processed into bootstrap commands that are executed during global cluster initdb (this happens at build time). Any new clusters created after the change will have the new row. For existing clusters that were created before the function was added, we need a migration script that manually inserts the pg_proc row. This script will get executed by ysql_upgrade.

The test works by comparing a freshly-created system catalog (generated locally by running initdb) to an old one which is downloaded from https://github.com/yugabyte/yugabyte-db-ysql-catalog-snapshots. There is a Mac and a Linux version of the snapshot; the Linux one was generated on AlmaLinux 8. Initdb loads the `en_US.utf8` collation from the OS libc and records some metadata, including the libc version, in the `pg_collation` table.

When someone is on another flavor of Linux, like Ubuntu, this test fails because of a libc version mismatch.

To fix this, we ignore the version of the `en_US.utf8` collation in this test. This is safe because we do not expect this collation to change between versions.
Jira: DB-15025

Test Plan:
Manual test (Alma 8):

Change `gnu_get_libc_version()` to `"2.39"` (or anything other than `"2.28"`) in this line:
```
~/code/yugabyte-db/src/postgres/src/backend/utils/adt/pg_locale.c:1716:
		/* Use the glibc version because we don't have anything better. */
		collversion = pstrdup(gnu_get_libc_version());
```

Then run the test:

```
./yb_build.sh release --java-test org.yb.pgsql.TestYsqlUpgrade#migratingIsEquivalentToReinitdb
```

To confirm that this test fails without this fix, make the changes to `pg_locale.c` but and run the old version of the test.

Reviewers: myang, mlillibridge

Reviewed By: myang, mlillibridge

Subscribers: mlillibridge

Differential Revision: https://phorge.dev.yugabyte.com/D41469