[#25801] DocDB: Remove txn reuse version field introduced for object locking feature
Summary:
For supporting object locks, earlier the idea was to version docdb transactions and reuse the same when the txn doesn't write any intents. The version felt necessary so as to prune obsolete wait-for edges at the deadlock detector since the only way they get pruned is when the status tablet responds that the txn is inactive/subtxn is inactive. Else, it could lead to false deadlocks.

Instead, if we ensure that we don't reuse a docdb readonly transaction when it ends up blocking a DDL, we wouldn't need transaction versioning at the first place. The wait-for edges on the transaction would get pruned once the transaction finishes.

This diff reverts the `txn_reuse_version` changes introduced as part of https://phorge.dev.yugabyte.com/D40202

**Upgrade/Downgrade safety**
1. Deprecating the field in productionized proto messages.
2. Deleting the field in object-locks specific proto messages since the usage is guarded by the test flag `enable_object_locking_for_table_locks`
Jira: DB-15101

Test Plan: Jenkins

Reviewers: amitanand, rthallam

Reviewed By: amitanand

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D41559