[#25803] docdb: Restart postgres when tserver restarts in minicluster tests
Summary:
We currently do not restart postgres when the associated tserver is restarted for our
(non-external) minicluster tests. Postgres and tserver share memory (TServerSharedData), and we end
up in a strange state not encountered in real clusters:
1. When MiniTabletServer is stopped, the underlying TabletServer is destroyed, which closes the TS
  file descriptor on the shared memory. At this point, postmaster still maintains its file
  descriptor, and is able to access the shared memory, so we don't encounter crashes, etc.
2. When the MiniTabletServer is restarted, a new TabletServer is created, which in turn creates a new
  TServerSharedData. This new object is a distinct object from the one PG processes are looking at,
  so new changes don't get propagated to the PG processes.

Furthermore, (1) depends on the postmaster having an open file descriptor to the shared memory to
keep it alive. While this is the case today, future changes to introduce shared memory allocators
will not maintain this constraint, resulting in backend processes failing to open shared memory.

This diff makes changes to also stop/start the postmaster in tandem with the associated tablet
server for minicluster tests.

PgMiniTestBase::PickPgTabletServer was also removed (misleading dead code) and PG startup code
was deduplicated between CDC tests and PostgresMiniCluster.
Jira: DB-15103

Test Plan: Jenkins

Reviewers: xCluster, hsunder, rthallam, telgersma, jason

Reviewed By: rthallam, telgersma, jason

Subscribers: jason, telgersma, rthallam, yql, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D41573