[PLAT-16630][dr] Create the safe time alert for new customers
Summary:
Currently, if you start with an empty YBA DB, the safe time alert won’t be automatically created as the customer table is empty and the migration script that is supposed to do that will not insert the corresponding row to the alert configuration table because there is no customer uuid to associate it with.

This adds `createForNewCustomer: true` to the safe time alert definition so that the alert is automatically created as part of YBA start up process.

Test Plan: Run the YBA on an empty database and made sure the alert is automatically created.

Reviewers: amalyshev, jmak

Reviewed By: amalyshev

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41584