[#24649] YSQL: enable INSERT ON CONFLICT batching for FK
Summary:
Previously, the presence of after row insert triggers prevented INSERT
ON CONFLICT read batching.  Loosen this restriction to allow for after
row insert triggers as long as they are for foreign key constraint
checking.  No further code changes should be needed because the current
design is sufficient.  Foreign key constraints are checked at the end of
the _statement_ or _transaction_, making them easy to reason with when
it comes to batching.
Jira: DB-13713

Test Plan:
On Almalinux 8:

    #!/usr/bin/env bash
    set -euo pipefail
    ./yb_build.sh fastdebug --gcc11
    find java/yb-pgsql/src/test/java/org/yb/pgsql -name 'TestPgRegressInsertOnConflict*' \
    | grep -oE 'TestPgRegress\w+' \
    | while read -r testname; do
      ./yb_build.sh fastdebug --gcc11 --java-test "$testname" --sj
    done

Close: #24649

Reviewers: kramanathan, #db-approvers

Reviewed By: kramanathan

Subscribers: svc_phabricator, yql

Differential Revision: https://phorge.dev.yugabyte.com/D41108