[BACKPORT 2024.2][#25817] DocDB: Add capability to FATAL a tserver from the master
Summary:
When a yb-tserver starts up it heartbeats with the master. At this time the master can determine the tserver is not allowed to join the universe and should immediately crash.
For example, if an older tservers tries to join a universe that has already been upgraded and finalized to a newer version. Check #25785/D41510 to see how this is used in the newer version.

**Upgrade/Downgrade safety:**
New proto field is optional and only set after the upgrade to a version higher than this version.

Fixes #25817
Jira: DB-15114

Test Plan: Manually tested with higher version

Reviewers: zdrudi, #db-approvers

Reviewed By: zdrudi, #db-approvers

Subscribers: slingam, svc_phabricator, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D41601