[PLAT-16611][PLAT-16625] Text update for failed node agent installation banner on YBA UI
Summary:
This diff updates the text on the failed node agent installation reminder banner on YBA UI.
The text update helps clarify the action that the user should take to resolve the failed installations.

This diff also updates the error labels for node agents. If the universe is paused, then the node agents under
that universe will be marked with `Universe Paused` if it has any error.

This diff also enables sorting for several columns that previously could not be sorted.

The `Delete` node agent action is renamed to `Unregister node agent` in order to maintain consistency with the term used in YBA documentation.
This action is no longer disabled for non-manually provisioned on-prem providers as we support this use case.

The `Reinstall node agent` action is now disabled for manually provisioned on-prem providers. This is because SSH access might not be available in this case and also any existing node-agent config will be wiped out when reinstalling.

The diff also changes the unregister node agent modal text to include the node name instead of the provider name as it is more relevant.

Test Plan:
Verify the text on the failed node agent installation banner is as expected.
{F329423}
{F329756}
{F329807}
{F329817}

Reviewers: ghodge, rmadhavan, kkannan, lsangappa, nsingh

Reviewed By: rmadhavan, nsingh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41528