[#24666] DocDB: Enable packed row for upgrade scenarios
Summary:
Enable packed row for upgrade scenarios. It is now safe to enable Packed row for upgrade scenarios as well as the issues around xCluster integration are not applicable, since most customers that rely on packed and xCluster are on 2.20+ builds.

Note that for customers on OSS and on builds lower than 2.20 and then upgrade to a build with this fix, the xCluster replication would pause and it needs to be re-established.

Test Plan: Jenkins

Reviewers: hsunder, slingam, sergei, #db-approvers

Reviewed By: hsunder, #db-approvers

Subscribers: svc_phabricator, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D41634