[#25854] YSQL: Extend existing backfill unit test to cover the index inconsistency bug #25250
Summary:
Commit ea73ed26 fixed the index inconsistency bug #25250 that occurs
during index backfill. A new test was added to cover it. There is an
existing unit test PgIndexBackfillTest.ReadTime to check for index consistency
when there are concurrent writes during backfill. However, it doesn't catch the
bug because of 2 reasons:

(1) the bug surfaces when there are backfill writes to multiple index
tablets. The test has only 1 index tablet.
(2) the test updates an index key and asserts for the presence of the
new key but misses to assert for the deletion of the old index key.

The test is modified to fix both issues.

The new test that was added in ea73ed26 is still retained because it is a
pointed test for this bug and actually does a concurrent delete of an index key
after index backfilling has started but before it backfills that same key. In
the PgIndexBackfillTest.ReadTime test the UPDATEs are issued after the backfill
read and write times are chosen but the backfill is blocked until
the UPDATEs finish.
Jira: DB-15148, DB-14442

Test Plan:
./yb_build.sh release --gtest_filter 'PgIndexBackfillTest.ReadTime'

Verified that the test fails without commit ea73ed26.

Reviewers: jason

Reviewed By: jason

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D41645