[PLAT-13783]Add new live query fields
Summary:
Add couple of new fields to the live query API response.
These fields arrived with the PG15 upgrade.

These fields were added in the yb-db as part of this diff: D41129

Test Plan:
Verified manually that the new fields are returned in the API response.
Tested the changes on pg11 and pg15 universes.

Reviewers: #yba-api-review!, vbansal

Reviewed By: vbansal

Subscribers: svc_phabricator, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41647