Merged PR 1566979: [OSS] Fix tests to run properly
### Does this PR have any customer impact?
No
### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Tests
### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No
### Are you introducing any new config? If yes, do you have tests with and without them being set?
No
### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description
As part of porting tests to OSS, some of the schemas were mangled and this caused the tests to test nothing and the baselines to be checked in incorrectly.

Fix those and add validation to prevent this specific type of error

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix for failing tests.

#### PR Summary
This pull request addresses issues with failing tests by correcting line references in the expected output files.
- `oss/internal/pg_documentdb_distributed/src/test/regress/expected/bson_query_operator_tests_index.out`: Updated line numbers in error messages to match the correct references in the SQL test files.