[#21160] YSQL: Fix update queries where pkey is set to self
Summary:
### The problem
Consider a simple table with a primary key, and no secondary indexes, constraints and triggers.
```
CREATE TABLE simple (k INT PRIMARY KEY, v INT);
INSERT INTO simple (SELECT i, i FROM generate_series(1, 10) AS i);
```
An UPDATE query where the primary key is set to itself produces the following error:
```
yugabyte=# UPDATE simple SET k = k, v = v + 1 WHERE k = 1;
ERROR:  42703: Missing column ybctid in DELETE request
```

### The fix
In most scenarios when the key columns of a tuple are updated, the query makes use of the distributed transaction path
in which the existing tuple is scanned into the plan slot. However, in a query where the key columns are set to themselves,
there currently exists a planner optimization which recognizes that the the columns are not being modified. If all the other conditions
for using a single shard transaction path are satisfied (as in the case of the example), then the single shard path is used
wherein the existing tuple is not fetched into the scan slot and the updated version of the tuple (in the slot) is then used to
construct the ybctid of the tuple to be deleted.
This revision disallows using the single shard transaction path if the query includes any of the columns making up the primary key
in its SET clause, even if the columns are set to themselves. This prevents the above query from using `YBCExecuteUpdate()` which
requires the existing tuple to have been fetched into the scan slot, thus side stepping the error. The error check still exists to guard
against incorrect use of the function.
Jira: DB-10094

Test Plan:
Run the following test:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressUpdateOptimized#schedule'
```

Reviewers: aagrawal, amartsinchyk

Reviewed By: aagrawal, amartsinchyk

Subscribers: smishra, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D41456