[#25872] docdb: Fix ReservedAddressSegment infinite loop in TSAN
Summary:
When multiple large reserved address segments are made in one process (notably, in the
case of minicluster tests, where multiple tservers/masters, and thus their reserved address
segments, are in the same process), it is possible we get into a situation where there is no block
available in the addresses we search for for TSAN (0x7e8000000000-0x7f6000000000).

In this case, we get a block farther away (0x55xxxxxxxxxx). If this is unacceptable to the child
process (and it is very likely to be, since it's located near the heap), then we get stuck:
we keep proposing an address in the 0x7xxxxxxxxxxx range, and get back the same block (since it's
the closest available one), and it keeps getting rejected.

This diff delays unmapping rejected blocks on the parent until after negotiation has finished, to
ensure that we always get back different blocks from mmap.

Additionally, a non-shutdown error status was sometimes returned on the parent when shutting down
negotiation; this has been fixed.
Jira: DB-15179

Test Plan:
Added tests:
```
./yb_build.sh --cxx_test 'util_reserved_address_segment-test' --gtest_filter 'ReservedAddressSegmentTest.TestNegotiationShutdown'
./yb_build.sh --cxx_test 'util_reserved_address_segment-test' --gtest_filter 'ReservedAddressSegmentTest.TestMultiple'
```
and ran 100x on TSAN.

Reviewers: sergei

Reviewed By: sergei

Subscribers: rthallam, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D41682