[#25879] YSQL: pg_upgrade Check for Username and cluster version
Summary:
User name must always be specified. Fail with a friendly error message if not specified.
Check for the cluster version and fail if it is not supported. This helps prevent user errors when the wrong version of the tool is used.

Fixes #25879
Jira: DB-15186

Test Plan: Pg15UpgradeTest.CheckVersion

Reviewers: nikhil, #db-approvers

Reviewed By: nikhil

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D41691