[#25855] YSQL: Remove sidecars from YBPgsqlLockOp
Summary:
Nowadays the `YBPgsqlLockOp` class inherits `YBPgsqlOp` which supports sidecars. But sidecars is not required by the `YBPgsqlLockOp`.
The main change in the diff is the introduction of the `YBPgsqlOpSidecarBase` intermediate base class which is inherited by the classes required sidecars.

Hierarchy prior to the change:

```
YBOperation
..YBPgsqlOp
....YBPgsqlLockOp
....YBPgsqlWriteOp
....YBPgsqlReadOp

```

Hierarchy after the change:

```
YBOperation
..YBPgsqlOp
....YBPgsqlLockOp
....YBPgsqlOpSidecarBase
......YBPgsqlWriteOp
......YBPgsqlReadOp
```

**Note:**
In context of this diff some code cleanup in related areas is performed
Jira: DB-15149

Test Plan:
Jenkins

Code cleanup was performed in existed unit test
```
./yb_build.sh --gtest_filter AdvisoryLockTest.*
```

Reviewers: sergei, qhu, bkolagani

Reviewed By: sergei, bkolagani

Subscribers: ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D41343