[PLAT-16668] Improving root check for yba-ctl preflight
Summary: Our root check wasn't properly handling preflight case where previous install did not have as_root set, so this diff removes the requirement for only handling no as_root when running certain commands, as if it's not set we are not getting good results from the config anyways and it is causing yba-ctl to error in confusing situations.

Test Plan: Upgrade from 2.20.4 to 2024.2.1 and ensure preflight checks now pass with change.

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41697