[PLAT-15267] Add additional xCluster warning banners
Summary:
This diff adds additional xCluster warning banners to increase awareness of tables which require attention.
The diff also adds a label `Some tables need attention` to the xCluster replication status and xCluster DR state
labels. This is done to highlight the existence of these concerning tables despite the config state/status appearing
as "health" with a green checkmark.

The config state is responsible for tracking the state of the replicaiton config itself and doesn't get set to
something unhealthy if a subset of tables happen to have schema mismatch with the target for example.

This diff aims to address this issue by surfacing the tables of concern through the added banners and new label next
to the config status.

Test Plan:
Verify the appropriate banners appear visible according to the current set of tables and replication errors.

xCluster replication:
{F328770}
{F328771}

xCluster dr:
{F331115}

Reviewers: rmadhavan, hzare, cwang, mkazerooni

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41506