[PLAT-16632] YBC Port customization
Summary:
Added a field "YB Controller RPC Port" in the Create Universe UI when the Override Deployment Ports toggle is enabled.
Wired it with the backend to override the port as necessary.

Test Plan: Checked whether the overridden port is correctly assigned in the nodes for the YB Controller RPC by ssh-ing into the node.

Reviewers: vkumar, kkannan

Reviewed By: kkannan

Subscribers: svc_phabricator, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41706