[PLAT-16403]Aborting a task on a universe from the task list page cancels a task on another universe
Summary:
Currently, we generate a <YBConfirmModal /> for all tasks in the running state and display it when VisibleModal is set to 'confirmAbortTask'. If multiple tasks are running, we render multiple modals, each overriding the previous one. This caused a bug where aborting a task would only abort the last task that overrode the previous ones.

Fix:
We have moved the <YBConfirmModal /> to an outer scope and introduced an internal state to store the task ID being aborted. This ensures that the correct task is aborted, preventing the issue.

we removed <TaskDetailsDrawer/> for now. We will fix this in the next diff.

Test Plan: Tested manually by creating multiple tasks and aborting them

Reviewers: lsangappa

Reviewed By: lsangappa

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41709