[#19445] [#25819] CDC: Introduce cdcsdk_flush_lag metric and fix other cdc metrics issues
Summary:
This diff has the following changes:

  # **Introduce a new lag metric called cdcsdk_flush_lag.**
This is a gauge metric and is applicable only for the logical replication model. Like other CDC metrics, this is also a tserver metric aggregated at stream level. For each {tablet, stream} pair this metric is calculated as the difference between the last replicated time in WAL and the restart time for the slot. Once the value of the metric is calculated for each {tablet, stream} pair, the max value among all the tablets for a stream is written to Prometheus. For gRPC streams where we don't have the concept of restart time, the value of this metric is always zero. The unit of this metric is microseconds.

  2. **Fix a bug in cdcsdk_sent_lag_micros metric**
There exists a bug in the calculation of cdcsdk_sent_lag_micros metric at the end of each GetChanges call because of which the value of the metric could potentially become negative when no workload is present. This bug in `CDCServiceImpl::UpdateTabletCDCSDKMetrics()` has been fixed by setting the metric value to max(0, calculated_value).

  3. **Stop counting the safepoint records in CDC throughput metrics**
Currently we include the safepoint records in the calculation of cdcsdk_change_event_count and cdcsdk_traffic_sent metric. This can lead to misleading metric values, when CDC is not shipping any valid DMLs / DDLs. In this diff we change this behaviour to only include valid DDLs and DMLs sent in GetChanges response for the calculation of cdc throughput metrics. To avoid iterating over the entire GetChanges response at the end for calculating these metric values, a pointer is passed to CDC producer. Every time a new record is added to the response we increment the values pointed to by these pointer. These values are then used to set the cdc throughput metrics.

  4. **Fix a bug in cdcsdk_traffic_sent metric calculation**
The cdcsdk_traffic_sent metric tells us the amount of data shipped by GetChanges in bytes. However the current calculation of this metric is faulty where we obtain its value by multiplying the number of records in response with the byte size of the first record in response. Since the first record is mostly a BEGIN record, this metric would end up under reporting the value. This has been fixed by the same mechanism described above in point 3.
Jira: DB-8236, DB-15116

Test Plan:
Jenkins: urgent
Jenkins: test regex: .*CDC.*

./yb_build.sh --cxx-test integration-tests_cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestFlushLagMetricWithRestartTimeMovement
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestCDCFlushLagMetricWithgRPCModel

Reviewers: xCluster, hsunder, siddharth.shah, skumar, vkushwaha

Reviewed By: siddharth.shah

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D41589